Move wormhole accounts to manager + add emitter
