+++ solana/programs/ntt-transceiver/src/wormhole/accounts.rs
@@ -1,37 +1,12 @@
 use anchor_lang::prelude::*;
+use example_native_token_transfers::wormhole_accounts::{pay_wormhole_fee, WormholeAccounts};
 use wormhole_anchor_sdk::wormhole;
 use wormhole_io::TypePrefixedPayload;
-
-// TODO: should we add emitter in here too?
-#[derive(Accounts)]
-pub struct WormholeAccounts<'info> {
-    // wormhole stuff
-    #[account(mut)]
-    /// CHECK: address will be checked by the wormhole core bridge
-    pub bridge: Account<'info, wormhole::BridgeData>,
-
-    #[account(mut)]
-    /// CHECK: account will be checked by the wormhole core bridge
-    pub fee_collector: UncheckedAccount<'info>,
-
-    #[account(mut)]
-    /// CHECK: account will be checked and maybe initialized by the wormhole core bridge
-    pub sequence: UncheckedAccount<'info>,
-
-    pub program: Program<'info, wormhole::program::Wormhole>,
-
-    pub system_program: Program<'info, System>,
-
-    // legacy
-    pub clock: Sysvar<'info, Clock>,
-    pub rent: Sysvar<'info, Rent>,
-}
 
 pub fn post_message<'info, A: TypePrefixedPayload>(
     wormhole: &WormholeAccounts<'info>,
     payer: AccountInfo<'info>,
     message: AccountInfo<'info>,
-    emitter: AccountInfo<'info>,
     emitter_bump: u8,
     payload: &A,
     additional_seeds: &[&[&[u8]]],
@@ -43,7 +18,7 @@ pub fn post_message<'info, A: TypePrefixedPayload>(
     let ix = wormhole::PostMessage {
         config: wormhole.bridge.to_account_info(),
         message,
-        emitter,
+        emitter: wormhole.emitter.to_account_info(),
         sequence: wormhole.sequence.to_account_info(),
         payer: payer.to_account_info(),
         fee_collector: wormhole.fee_collector.to_account_info(),
@@ -63,26 +38,6 @@ pub fn post_message<'info, A: TypePrefixedPayload>(
         TypePrefixedPayload::to_vec_payload(payload),
         wormhole::Finality::Finalized,
     )?;
-
-    Ok(())
-}
-
-fn pay_wormhole_fee<'info>(
-    wormhole: &WormholeAccounts<'info>,
-    payer: &AccountInfo<'info>,
-) -> Result<()> {
-    if wormhole.bridge.fee() > 0 {
-        anchor_lang::system_program::transfer(
-            CpiContext::new(
-                wormhole.system_program.to_account_info(),
-                anchor_lang::system_program::Transfer {
-                    from: payer.to_account_info(),
-                    to: wormhole.fee_collector.to_account_info(),
-                },
-            ),
-            wormhole.bridge.fee(),
-        )?;
-    }
 
     Ok(())
 }
